Fix CSS proxy
