+++ projects/plugins/boost/app/features/optimizations/critical-css/Critical_CSS.php
@@ -64,6 +64,8 @@ class Critical_CSS implements Feature, Has_Endpoints {
 		}
 
 		Critical_CSS_Invalidator::init();
+		CSS_Proxy::init();
+
 		add_filter( 'jetpack_boost_js_constants', array( $this, 'add_critical_css_constants' ) );
 		add_filter( 'jetpack_boost_admin_notices', array( $this, 'add_admin_notices' ) );
